requirements.txt
Django==5.2.9
djangorestframework==3.19.1
requests==2.31.0
geopy==2.4.0
python-dotenv==1.1.0
polyline==1.5.0

